pyproject.toml
[tool.poetry]
name = "buspi"
version = "0.1.0"
description = ""
authors = ["Nico Westerbeck <Nico.Westerbeck@mailbox.tu-dresden.de>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
requests = "^2.32.3"
rpi-gpio = "^0.7.1"
python-dateutil = "^2.9.0.post0"


[tool.poetry.group.dev.dependencies]
jupyter = "^1.1.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
